Added functionality to the first two buttons for all the Computer Science modules. When one of these buttons are clicked, the corresponding resource file (.pdf) will be downloaded on the user's phone and the user will be able to access and view it. The buttons have also been renamed to the name of the resource and what type of resource it is so the user knows what they are downloading.
Added functionality to the first two buttons for all the Computer Science modules. When one of these buttons are clicked, the corresponding resource file (.pdf) will be downloaded on the user's phone and the user will be able to access and view it. The buttons have also been renamed to the name of the resource and what type of resource it is so the user knows what they are downloading.