Added functionality to the last two buttons for all the Computer Science modules. When one of these buttons are clicked, the corresponding resource website will be loaded within the app and the user will be able to view and browse the website. The buttons have also been renamed to the name of the resource and what type of resource it is, so the user knows what website they are visiting. A ‘go back’ button has also been added so when the user opens a website resource, the user can click this button which will close the website within the app and show all the available resources again to the user.
Added functionality to the last two buttons for all the Computer Science modules. When one of these buttons are clicked, the corresponding resource website will be loaded within the app and the user will be able to view and browse the website. The buttons have also been renamed to the name of the resource and what type of resource it is, so the user knows what website they are visiting. A ‘go back’ button has also been added so when the user opens a website resource, the user can click this button which will close the website within the app and show all the available resources again to the user.