Grade calculator screen now will not crash when submitting nothing. A toast will also pop up with an error to let the user know that it did not work.
Submitting your grades in the modules screen will now show a toast to the user that it was successful, otherwise an error it if was not.

Changes made to the toolbars to add a refresh button so that when module grades are updated, the user can refresh the screen for up to date information.