Completed backend for the assessment resources screen using Firebase. When any of the resource buttons are clicked, the corresponding resource file will be downloaded to the user's mobile phone. A back button has also been added on the assessment resources screen which allows the user to return the resources screen when it is clicked.
Completed backend for the assessment resources screen using Firebase. When any of the resource buttons are clicked, the corresponding resource file will be downloaded to the user's mobile phone. A back button has also been added on the assessment resources screen which allows the user to return the resources screen when it is clicked.