Created the screen for 'Request a Resource', the front end for this screen is completed. A toolbar has been added on both the resources screen as well as the assessment resources screen. In this new toolbar, the user can open up the menu by clicking icon where they can choose to view the 'Request a Resource' screen.
Created the screen for 'Request a Resource', the front end for this screen is completed. A toolbar has been added on both the resources screen as well as the assessment resources screen. In this new toolbar, the user can open up the menu by clicking icon where they can choose to view the 'Request a Resource' screen.