Completed the backend for the 'Request a Resource' screen using SQL. On this screen, there are two dropdown lists where the user can select one of the modules (CS2001, CS2002, CS2003, CS2004, CS2005) that the resource is related to as well as the type of resource it is (book, website, assessment, other). Then, there is an input field for the name of resource which can be entered by the user, once that is done the user can click the submit button which will insert the data submitted by the user into the SQL database and table.
Completed the backend for the 'Request a Resource' screen using SQL. On this screen, there are two dropdown lists where the user can select one of the modules (CS2001, CS2002, CS2003, CS2004, CS2005) that the resource is related to as well as the type of resource it is (book, website, assessment, other). Then, there is an input field for the name of resource which can be entered by the user, once that is done the user can click the submit button which will insert the data submitted by the user into the SQL database and table.