Updated the colour of the dropdown menus, input field, buttons and background of the 'Request a Resource' screen to match the rest of the application's colour scheme. Also, added a toast message which lets the user know that their resource request has been submitted when the user clicks the 'Submit Resource' button.
Updated the colour of the dropdown menus, input field, buttons and background of the 'Request a Resource' screen to match the rest of the application's colour scheme. Also, added a toast message which lets the user know that their resource request has been submitted when the user clicks the 'Submit Resource' button.